Renamed all instances of Activity to Workout
